Added JGrapht as a dependency
Need to create a parser that will figure out how to create the graph
from the given bionetlist. This can be further used to create a
BioFunction in Fluigi by just mapping the graph created here to the
fluid graph in Fluigi. We then use the existing infrastructure to
generate outputs.